+++ src/AlfBasicClass.php
@@ -9,21 +9,17 @@ use JetBrains\PhpStorm\ArrayShape;
 
 abstract class AlfBasicClass {
 
+    /** @AlfAttrAutoComplete */
+    #[AlfAttrAutoComplete]
+    final public static function _AlfBasicClass($obj) : AlfBasicClass {
+        return AlfProgramming::_()->unused($obj);
+    }
+
     public function __construct() {
         foreach (array_reverse($this->listPhpTraits()) as $traitName) {
             $callCTorFunc = '_'.$traitName.'CTor';
             $this->$callCTorFunc();
         }
-    }
-
-    public function listPhpTraits() : array {
-        return AlfPhpClassManager::_()->listTraits($this);
-    }
-
-    /** @AlfAttrAutoComplete */
-    #[AlfAttrAutoComplete]
-    final public static function _AlfBasicClass($obj) : AlfBasicClass {
-        return AlfProgramming::_()->unused($obj);
     }
 
     public function __destruct() {
@@ -33,20 +29,27 @@ abstract class AlfBasicClass {
         }
     }
 
-    // TODO: public function __wakeup() : void
-    // TODO: public function __sleep() : array
-    // TODO: public function __serialize() : array
-    // TODO: public function __unserialize(array $data) : void
-
     public function __clone() {
         foreach (array_reverse($this->listPhpTraits()) as $traitName) {
             $callCloneFunc = '_'.$traitName.'Clone';
             $this->$callCloneFunc();
         }
     }
+    // TODO: public function __wakeup() : void
+    // TODO: public function __sleep() : array
+    // TODO: public function __serialize() : array
+    // TODO: public function __unserialize(array $data) : void
+
+    public function getPhpParentClass() : ?string {
+        return AlfPhpClassManager::_()->getParent($this);
+    }
 
     public function listPhpParentClasses() : array {
         return AlfPhpClassManager::_()->listParents($this);
+    }
+
+    public function listPhpTraits() : array {
+        return AlfPhpClassManager::_()->listTraits($this);
     }
 
     #[ArrayShape(['class' => "string", 'parent' => "string"])]
@@ -55,10 +58,6 @@ abstract class AlfBasicClass {
             'class'  => static::class,
             'parent' => $this->getPhpParentClass(),
         ];
-    }
-
-    public function getPhpParentClass() : ?string {
-        return AlfPhpClassManager::_()->getParent($this);
     }
 
 }