+++ src/AlfBasicTypeScalar.php
@@ -3,13 +3,17 @@
 namespace Alf;
 
 use Alf\Attributes\AlfAttrAutoComplete;
+use Alf\Interfaces\Values\AlfNullWork;
+use Alf\Interfaces\Values\AlfNullWorkTrait;
 use Alf\Services\AlfProgramming;
 
-abstract class AlfBasicTypeScalar extends AlfBasicType {
+abstract class AlfBasicTypeScalar extends AlfBasicType implements AlfNullWork {
+
+    use AlfNullWorkTrait;
 
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
-    public static function _AlfBasicTypeScalar($obj) : AlfBasicTypeScalar {
+    final public static function _AlfBasicTypeScalar($obj) : AlfBasicTypeScalar {
         return AlfProgramming::_()->unused($obj, static::_AlfBasicType($obj));
     }