+++ src/Types/Scalars/AlfIntRange.php
@@ -10,12 +10,12 @@ abstract class AlfIntRange extends AlfInt {
 
     /** @AlfAttrAutoComplete */
     #[AlfAttrAutoComplete]
-    public static function _AlfIntRange($obj) : AlfIntRange {
+    final public static function _AlfIntRange($obj) : AlfIntRange {
         return AlfProgramming::_()->unused($obj, static::_AlfInt($obj));
     }
 
     #[Pure]
-    protected function convertValueForSet(int $value) : ?int {
+    protected function _convertValueForSet(int $value) : ?int {
         if ($value > $this->getValueMax()) {
             return $this->getValueMax();
         }