+++ tests/Basic/ClassTest.php
@@ -5,7 +5,11 @@ use Alf\AlfBasicClass;
 test('class must be a instance of AlfBasicClass',
     /** @throws ReflectionException */
     function (string $className) {
-        $reflection = new ReflectionClass($className);
-        $isAlfBasicClass = $reflection->isSubclassOf(AlfBasicClass::class) || ($className === AlfBasicClass::class);
+
+        $reflectionClass = new ReflectionClass($className);
+
+        // -
+        $isAlfBasicClass = $reflectionClass->isSubclassOf(AlfBasicClass::class) || ($className === AlfBasicClass::class);
         expect($isAlfBasicClass)->toBeTrue();
+
     })->with(listAlfClasses());